Bat & Aliens Animations & Setup
I added the animation to the bat and is now complete. There was no Alien models so I went ahead and added some, I slightly set them up, like adding rigidbodies and colliders, but I would like to leave the tweaking to you guys and those animations are done too.